Improve the performance of the message viewer API (#27)
* chore: Convert the object mapper to a constant variable

* fix: Improve the performance of the message viewer API

The message viewer API consumes raw records from Confluent Cloud,
deserializes the records if they use a schema, and
returns the processed records.

Prior to this change, the message viewer API used Java's
[List.indexOf()](https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/util/List.html#indexOf(java.lang.Object))
to look up the index of a record in the response object
after applying the schema and before updating it. `List.indexOf`
uses a sequential search and is not very performant
when dealing with larger lists of larger `record`s.

While the JVM's JIT compiler is able to optimize this code
over time, GraalVM's AOT compiler is not, leading to a
poor performance of the message viewer API in the native
executables.

This change updates the processing and updating of records
to use the Java Streams API so that we update records
in-place. On my machine, this change reduces the time needed for
processing/updating records from ~ 700ms to less than 5ms, i.e.,
it makes every request against the message viewer API almost
one second faster.