Always remove cached schema registry clients and failed schema errors when connection disconnects
When any connection is closed, remove all Schema Registry clients that were cached for that connection.

It also clears all cached errors from failing to load schemas for any connection. Unfortunately, those are not cached by connection or Schema Registry client, so that clearing will have to be improved in a future PR.

For what it’s worth, we only use Schema Registry with CCloud connections. We plan to add SR usage for local connections, and when Confluent Platform is supported.