feat: Make settings related to SSL certificates configurable via Preferences API (#24) (#64)
* feat: Make settings related to SSL certificates configurable via Preferences API

A few users of [Confluent for VS Code](https://github.com/confluentinc/vscode) (https://github.com/confluentinc/vscode/issues/280 https://github.com/confluentinc/vscode/issues/268) reported that they ran into errors when working with Confluent Cloud and sitting behind a firewall that performs SSL inspection.

This change introduces the Preferences API that can be used to store user preferences in the sidecar.

In the beginning, it supports two preferences related to SSL handshaking:
* The preference `tls_pem_paths` can be used to point the sidecar to one or multiple custom certificates in the PEM format (by default, no custom certificate is configured).
* The preference `trust_all_certificates` can be used to enable/disable the verification of the server's certificate (by default, verification is enabled).

Whenever the preferences change, a [CDI event](https://quarkus.io/guides/cdi#events-and-observers) gets produced so that beans relying on the preferences
get notified. This change updates the `WebClientFactory` to `@Observe` the events related to the Preferences API so that
the configuration of the web client is updated whenever user change their preferences. In the future, other beans consuming
the preferences may `@Observe` the events, too.

Fixes #63

* chore: Remove empty lines

* chore: Naming

* test: Verify that updating the preferences changes the web client config

* chore: Use var consistently