Initial support for using LaunchDarkly feature flags
This adds the infrastructure for using LaunchDarkly feature flags and exposing a new REST API to allow the VS Code extension to _list_ feature flags, _read_ a specific feature flag (including the current evaluated value of the flag), and get just the currently evaluated value of a specific feature flag.

LaunchDarkly does not yet have a Java client SDK that uses a non-secret client ID. (There is a Java server SDK, but this is intended as a multitenant services and requires an API key and secret that should not be shared.)

This framework uses a simple HTTP client to poll the HTTP API to get the currently evaluated flags given a context. The flags are evaluated on a schedule (e.g., every 60 seconds) and cached. If the sidecar is unable to evaluate the flags the first time (e.g., LaunchDarkly is not available when the sidecar starts), a set of default flags is used. However, as soon as flags can be evaluated the results are cached. Subsequent failures to refresh the flags (e.g., LaunchDarkly goes down after the sidecar process was started) will simply keep the cached flags (there is no TTL).

This framework also allows code within the sidecar to easily check feature flags. The `FeatureFlag` methods to get feature flag values require `FlagId` enum literals, and unit tests ensure that there is a default value for all `FlagId` literals. This means there is always a value for these flags, and centralizes the default values in a JSON file on the classpath (the same format that is returned from LaunchDarkly).

Because the sidecar version is needed in the device context, this also refactors how that information is kept (moving it out of the `VersionResource` and into a new `SidecarInfo` application bean with several other attributes. This bean and the various attributes can be injected into other beans, including the new `FeatureFlags`.

Currently the OS information is read from the System properties, and the VS Code version and extension version information is read from either System properties (lowercase dot-delimited), such as
```
ide-sidecar-*-runner -Dvscode.version=1.93.2 -Dvscode.extension.version=v0.17.1 …
```
or if those are absent from environment variables (capital case underscore-delimited) used to start the sidecar process, such as:
```
VSCODE_VERSION=1.93.1 VSCODE_EXTENSION_VERSION=v0.17.0 ide-sidecar-*-runner …
```
These are currently optional, but we do want to change the extension to start supplying them soon, as the VS Code information is passed in the `device` contexts passed to LaunchDarkly during flag evaluation.