Incorporate reviewer suggestions
Refactored the `FeatureFlagsFailedException` and how the LD error response is parsed and that exception object thrown.

Also changed the `WebClientFactory` fields to use an `@Inject`-ed instance. Since there are few beans in the `featureflags` package, this change required passing the `WebClientFactory` into the other objects in a different way. In particular, the `FeatureProject` objects are not beans and are instead instantied statically, before the `WebClientFactory` field could be injected. So now the `WebClientFactory` is `@Inject`-ed into the `FeatureFlags` bean, and passed into the `FeatureProject.refresh(…)` method, which is where the client factory is used.

 These changes (particular the second) required quite a few simple changes to several test cases.