Made sure the `FeatureFlags` and `SidecarInfo` beans are initialized at application startup
With the `@Startup` annotation, the `FeatureFlags` and `SidecarInfo` beans are initialized when the sidecar starts up. The LD flags are evaluated immediately and likely ready before the flags are used in the sidecar or via the REST API.

Without that annotation, those beans are initialized only when they are used. In the case of `FeatureFlags`, that delayed initialization would be too late after the first feature flag was requested, as the evaluation of the feature flags would likely not yet have completed.

For example, if a REST request were for a feature flag with a default, the default would have been used. But if the request were for a feature flag that does not have a default (i.e., no `FlagId` enum literal, as is the case for all CCloud flags), the REST request response would have been a `404 Not Found` signifying the feature flag does not exist.