FeatureFlagFailureException is not exposed to API clients
Improved the handling of `FeatureFlagFailureException`. The `HttpFlagEvaluationProvider` emits it when it is unable to parse the LD evaluation response or an error response. Since the feature flags are always evaluated on a separate thread, while local checks of feature flags can occur on any thread and only hit the cached values.

Therefore, `FeatureFlagFailureException` is never exposed to users, and doesn’t need to be in the `ExceptionMappers`.