feat: Introduce new authentication status `FAILED`
This change introduces the property `isTransient` to the `AuthErrors`,
which models if an error is transient or not. It also introduces a new
`ConnectionStatus.Status`, `FAILED`, which is returned if a connection
hols at least one error that is marked as non-transient. We assume
that connections cannot recover from non-transient errors and advise the
extension to trigger a new authentication flow if the connection reports
the state `FAILED`. The sidecar does not attempt to refresh tokens of a
connection that holds a non-transient error.

We determine if an error is transient as follows:
* any error related to the `authStatusCheck` is considered to be transient
* any error related to the `signIn` is considered to be non-transient
* errors related to the `tokenRefresh` are considered to be non-transient if more than `CCloudOAuthConfig.MAX_TOKEN_REFRESH_ATTEMPTS` token refresh attempts have failed or if the CCloud API has reported that the refresh token is invalid or expired; otherwise, errors related to the `tokenRefresh` are considered to be transient

Example authentication status of a connection for which the CCloud API has reported that the refresh token is invalid or expired:

```shell
$ curl http://localhost:26636/gateway/v1/connections/1

{
  "api_version": "gateway/v1",
  "kind": "Connection",
  "id": "1",
  "metadata": {
    "self": "http://localhost:26636/gateway/v1/connections/1",
    "resource_name": null,
    "sign_in_uri": "https://login.confluent.io/oauth/authorize?..."
  },
  "spec": {
    "id": "1",
    "name": "1",
    "type": "CCLOUD"
  },
  "status": {
    "authentication": {
      "status": "FAILED",
      "errors": {
        "token_refresh": {
          "message": "Retrieving the ID token failed for the following reason: invalid_grant - Unknown or invalid refresh token.",
          "created_at": "2024-10-17T13:42:46.469860Z",
          "is_transient": false
        }
      }
    }
  }
}
```

Fixes #86