Initial support for direct connections
Fixes #96

This adds the initial support for a new type of connection resources, called `DIRECT` connections, that represent a local or remote Kafka cluster and/or Schema Registry. The extension might create a direct connection to talk to a locally running CP installation, or a remote Apache Kafka cluster with a remote Schema Registry, or even a CCloud cluster with Schema Registry. The user may have to provide credentials to successfully authenticate and connect to each remote service/cluster.

The direction connections are not necessarily the constructs we want to expose. It is likely that near-term versions of the VS Code extension will allow users to connect to local or remote Kafka and SR clusters, but it may also expose to users different kinds of connections that are all represented in the sidecar as direct connections. For example, maybe the VS Code extension allows users to connect to “Confluent Platform”, “Apache Kafka”, and “WarpStream” (in addition to CCloud and local AK+SR), and these might all be customized facades on top of direct connections.

This is the first step in full support for direct connections, and as such only limited functionality is available in this PR:
* CRUDL connection resources with `type=DIRECT`, with validation of fields for each type of connection.
* A direct connection contains zero or one `kafka_cluster` object and zero or one `schema_registry` objects, allowing it to connect to a (remote or local) Kafka cluster and/or a (remote or local) Schema Registry.
* GraphQL support for the Kafka cluster and SR referenced by direct connections.

The following changes will be made to this PR before it is to be merged, likely through subsequent PRs based on this PR:
# Add basic credentials to the `kafka_cluster` and `schema_registry` definitions, and use this to connect to downstream resources. All secrets in credentials will be write-only: the Connection REST API endpoins must always mask secrets.
# Update direct connection status by showing authn success against Kafka cluster and/or SR. This may mean we evolve the current `status` section of the connection resources to better work with non-CCloud connections, where connectivity to Kafka cluster and SR are independent.
# Support using direction connections in Kafka REST API and Message Consumer API endpoints
# Support using direction connections in SR API endpoints
# Local connections can define SR URL via same `schema_registry` field as direct, superseding the existing `local_config`, which will still be supported for a while so that the extension can migrate. A local connection is invalid if it has both `local_config` and `schema_registry` objects.

A few enhancements may come after this initial functionality is merged:
# Add an `PLATFORM` (or `MDS`) connection type that connects to a local or remote MDS and dynamically discovers the registered clusters.
# Potentially add custom client properties to `kafka_cluster` and `schema_registry`, allowing users to override some client connection properties.
# Improved validation logic for resource types, potentially using javax.validation.
# Allow direct connections to connect to Flink cluster and Kafka Connect cluster.