Refactor the sidecar integration tests to support different types of containers
Previously, we had a `ConfluentLocalTestBed` class that managed containers for Kafka and SR, and had utility functions (and state) for working with those containers and the sidecar process being tested.

This design had a few issues. First, it mixed together multiple things. Second, it required classes extend the `ConfluentLocalTestBed` class that would have made it difficult to use a different environment, say for integration testing. Third, it made it difficult to run a suite of tests on local connections AND on direct connections. Relying primarily upon inheritance is restrictive in Java due to the single base class approach.

This PR refactors the `ConfluentLocalTestBed` into two distinct components:
* a `LocalTestEnvironment` that extends a new`TestEnvironment` interface, used to manage containers for a Kafka cluster and Schema Registry cluster
* a `SidecarClient` that provides useful functions for working with a sidecar process that is being tested.

Instead of tests extending the `ConfluentLocalTestBed`, this PR introduces a new `AbstractSidecarIT` class that extends the `SidecarClient` to make the methods to operate on the sidecar more easily accessed by tests. More importantly related to local and direct connections, there is a set of connection-initialization methods that test classes can use in their `@BeforeEach` methods. This will make it far easier to extend existing IT classes to run the tests against direct connections.