Add GraphQL query support for direct connections (#133)
The `ConnectionList` and `TemplateList` did not fully support deserializing, since production code doesn’t require it. However, this is needed when using REST Easy to deserialize these API responses with:
```
var connectionsList = given()
        .when()
        .get()
        .then()
        .statusCode(200)
        .contentType(ContentType.JSON)
        .extract().body()as(ConnectionsList.class);
```
Prior to this change, most of our tests extracted the body as a JSON representation, then parsed it into `JsonNode`, and manipulated the result. But extracting the body into its deserialized form is far simpler.

This PR does not otherwise change API behavior or the OpenAPI specification, and only changes how the Connection API and Template API resource types are implemented and tested. Specifically, it refactors the `BaseList`, `BaseModel` classes and their subtypes to be more conventional (e.g., storing the serialized fields, rather than deriving serialized fields transiently from other hidden fields) and adds new creator methods and additional annotations. The PR changes a few tests to use the type-specific deserialization approach, and adds new tests to verify the types could be deserialized, serialized, and then deserialized again without losing any information. These changes make it very easy to add in the future new tests that use the type-specific deserialization.

This PR does not rewrite the numerous tests that use the JSON approach.

* Feedback from reviewers, discovered missing coverage and fixes required

* Add GraphQL query support for direct connections

The graphQL functionality for reading cluster resources was not quite working for direct connections. This adds that support and adds some unit tests to verify the queries function.

Limitations
* Cluster information returned in the GraphQL query results do not yet reflect whether the user has connectivity with or the ability to authenticate with the Kafka and Schema Registry clusters referenced in direct connections. Connection status will be added soon.
* Kafka REST and SR APIs also are not yet working with direct connections, and this will be fixed soon.