Change the `AbstractSidecarIT` base class to reuse the same containers and connections
Starting containers is time consuming and can take up to 5-10 seconds. Prior to this change, every `AbstractSidecarIT` would start its own containers and shut them down after all tests were run. Considering we already have 6 subclasses, this is almost a minute just to *restart* and shutdown the extra containers. And will likely have more subclasses in the future.

Also, all of the tests in those IT test classes ended up creating a new local connection, which could add 3-5 seconds to each test. And soon we’ll double the tests by running those tests against direct connections.

This PR changes the `AbstractSidecarIT` base class to efficiently reuse containers and connections across many tests. First, the `AbstractSidecarIT` base class starts up the `LocalTestEnvironment` containers before the first integration test is run, and then shut them down after *all* integration tests have completed. This is true no matter how many subclasses of `AbstractSidecarIT` there are. Oh, and all of the subjects and topics created in one tests are cleaned up between each test.

Second, this PR introduces the concept of a “test scope” to the creation of connections, so that connections can be created once and used for all of the integration tests that use the same scope. All of the `AbstractSidecarIT` subclasses use a test scope named for the concrete IT class, though the technique is flexible to allow other patterns. This change can save up to five minutes across a hundred tests.

Note that other `TestEnvironment` instances can be added to `AbstractSidecarIT` in the future.