feat: Consider baked-in default certs when using PEM file
When using a PEM file to verify the server-side TLS certs, make sure
that we do not lose access to the default certs baked into the native
executable.

We do not expect this change to cause any harm because we won't skip
any certs but consider a few additional ones.

Fixes #523 https://github.com/confluentinc/vscode/issues/523