Add native support for Kafka client auth mechanisms
Kafka clients and SR clients use LoginModule and CallbackHandlers for authentication, and these are loaded at runtime. Without proper registration of these classes, trying to configure a client that uses credentials in the native sidecar executable will result in runtime errors.

This PR registers a few more classes for reflection so they are included in the native image. It also relies upon the Quarkus extensions for Kafka client and SR client (w/ Avro) that have build item hooks, and which cover the majority of the registrations. A few LoginModule and CallbackHandler implementations available to our codebase are not defined in the Quarkus extensions, so we add all of the available implementation in the `@RegisterForReflection` annotation on `ReflectionConfiguration`.

More testing is still required.