Evolve the schema of the Connection resource `status`
Change the `status` object in `Connection` resources to able to represent the statuses of the different external clusters/resources that connections can define.

The `ConnectionStatus` OpenAPI spec schema now looks like:
```
    ConnectionStatus:
      required:
      - authentication
      type: object
      properties:
        ccloud:
          $ref: "#/components/schemas/CCloudStatus"
        kafka_cluster:
          $ref: "#/components/schemas/KafkaClusterStatus"
        schema_registry:
          $ref: "#/components/schemas/SchemaRegistryStatus"
        authentication:
          $ref: "#/components/schemas/Authentication"
```
The `status.authentication` object is unchanged, but is being deprecated and replaced with the `status.ccloud` object. The `status.ccloud`, `status.kafka_clsuter` and `status.schema_registry` nested objects all use the `ConnectedState` enum:
```
    ConnectedState:
      enum:
      - NONE
      - ATTEMPTING
      - SUCCESS
      - EXPIRED
      - FAILED
      type: string
```
and there is a direct mapping from the older `Status` enumeration to this new enum.

The `status.authentication` object is being generated from the `status.ccloud` object, and the sidecar code has already transitioned to only using `status.ccloud`; the `status.kafka_cluster` and `status.schema_registry` objects will be set in a subsequent PR.

We will keep the `status.authentication` object for a time being, so the extension has time to migrate to the `status.ccloud`, `status.kafka_cluster` and `status.schema_registry` objects. Once that migration has been performed, we will remote the `status.authentication` object altogether, in a future PR.