Make Consumer and Producer client config configurable (#167)
* feat: Make Consumer and Producer client config configurable

This change makes the configuration of the Kafka Consumer and Producer clients
configurable. Also, similar to the Kafka Admin clients, it sets the
`client.id` to a sidecar-specific value so that users can distinguish
between consume/produce requests of the ide-sidecar and those of other
Kafka applications.

* chore: Set defaults for configs `vscode.version` and `vscode.extension.version`

By default, set the configs `vscode.version` to unknown and `vscode.extension.version` to unknown.