Remove the `kafka_cluster.id` from the connection specs, and update RealDirectFetcher to populate the cluster ID
We should not allow users to specify the ID of the Kafka cluster in direct connection specs — getting the ID is not always straightforward, and even if they provide an ID it might be wrong.

Instead, the `RealDirectFetcher` should be able to use the AdminClient to get the cluster’s ID from the cluster, and then include the ID in the cached `KafkaCluster` objects.

Added integration tests for the Connection API and GraphQL API to test local and direct connections and the Kafka cluster and Schema Registry cluster representations (including the Kafka cluster does have an ID).
(Unfortunately, we can’t easily mock Kafka protocol, so adding unit tests for all the `RealDirectFetcher` logic is not as straightforward.)

Found and fixed a bug where the `status` response in a newly-created direct connection had no `kafka_cluster` or `schema_registry` object.