Removed the unit tests for DirectQueryResourceTest
The `RealDirectFetcher` class now uses AdminClient to verify the cluster specified in the connection is available. Since we’re not able to mock the Kafka native protocol, these tests are better performed with integration tests. (We could mock some subset of functionality in `RealDirectFetcher`, but it’s not the most complicated portion and is therefore considered not worthwhile.)