Support Connection API for testing possible connections
Resolves #179

Expands the `POST /gateway/v1/connections` endpoint to add a `dry_run` query parameter that defaults to `false` for current behavior. When `dry_run=true`, the endpoint will not create a new connection but will instead return the Connection resource type response with a status that contains the information _as if_ the connection were created. This will allow the extension to have a “Test” button on the new connection panel, without having to actually create, wait for updated status, and then delete the connection.

The behavior of the existing endpoint will be changed as follows:
* When the `dry_run` query parameter is `false` (the default), the endpoint calls `ConnectionStateManager.createConnectionState(spec)` that checks for duplicate IDs (if provided), validates the input spec (failing if there are validation errors), adds the new connection in the `ConnectionStateManager`, and returns a Connection object with the supplied `ConnectionSpec` and _initial_ `ConnectionStatus`.
* When the `dry_run` query parameter is `true`, the endpoint calls a new method `ConnectionStateManager.testConnection(spec)` that checks for duplicate IDs (if provided), validates the input spec (failing if there are validation errors), and *synchronously* computes the status (via `ConnectionState.getConnectionStatus()`), and finally returns a Connection object with the supplied `ConnectionSpec` and computed `ConnectionStatus`.

Found and fixed a bug in the `DirectConnectionState` that was incorrectly setting the `status.schema_registry.state` rather than null when there is no schema registry configured.