GraphQL queries should not block for direct connections that are not connected
Resolves #197. Also relates to #181

Modifies `ConnectionState` as follows:
* add a new `cachedStatus` field to store the most recent status
* renames the asynchronous `getConnectionStatus(): Future<ConnectionStatus>` method to `refreshConnectionStatus(): Future<ConnectionStatus>`, changed the implementation to:
  * delegate to a new protected method called `doRefreshConnectionStatus()` that subclasses override
  * update the `cachedStatus` field
* add a new *synchronous* `getConnectionStatus(): ConnectionStatus` method that returns the `cachedStatus` value

All the uses of the old async `getConnectionStatus(): Future<ConnectionStatus>` have been changed to use the newer async `refreshConnectionStatus(): Future<ConnectionStatus>` method.

The `DirectConnectionState` class was changed to have two convenience methods:
* `isKafkaConnected()` returns true when `status.kafkaCluster().isConnected()` is true; and
* `isSchemaRegistryConnected()` returns true when `status.kafkaCluster().isConnected()` is true

When we want to implement the async status update mechanism (#181), that mechanism only has to call the async `refreshConnectionStatus()`, which will then automatically cache the status in the `ConnectionState` instance.

This commit also changes `RealDirectFetcher` to use the two convenience methods to check the status of the connection’s Kafka cluster and SR *before* using the AdminClient or SR client, respectively, to construct the corresponding cluster information. Only when the status is “SUCCESS” will the clients be used; in all other cases, the `RealDirectFetcher` quickly returning no cluster information.

Added new unit tests for `RealDirectFetcher` to verify the behavior for various statuses and error conditions.